configuration/scripts/requirements.txt
ipython
matplotlib
mysql_connector_repackaged
pandas
plotly
python-dotenv
requests
sqlalchemy
kaleido
mysql-connector-python
pymysql